style(rtc_types::enclave_messages): extract mod set_access_key to file
